taken the MixedSegmentError from charmm.build, implemented _checkMixedSegment on builder, which both charmm and amber builds can use
